requirements.txt
flask
python-dotenv
pymongo[srv]
ariadne
ollama
gunicorn
streamlit
requests
pandas